Achievement 2/Exercise2.5/src/requirements/dev.txt
-r base.txt
ipython==8.18.1
django-extensions==3.2.3